requirements.txt
kafka-python
peewee
requests
pixivpy
loguru
pymysql